archived/ml-prediction/requirements.txt
# Enhanced ML Prediction Service Dependencies v2.1.0 - Advanced AI/ML Stack
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Enhanced Pydantic schemas with advanced validation
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic-extra-types==2.5.0

# HTTP client
httpx==0.27.0
aiohttp==3.9.1

# ML and Data Science
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
statsmodels==0.14.1
ta==0.11.0

# Portfolio Optimization
cvxpy==1.4.2
cvxopt==1.3.2

# LLM Integration
openai==1.12.0
ollama==0.2.1
llama-cpp-python==0.2.55
transformers==4.37.2
torch==2.2.0
tokenizers==0.15.1

# Caching and Storage
redis==5.0.1
aioredis==2.0.1

# Security and Privacy
cryptography==42.0.2
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Colombian Market Integration
requests==2.31.0
zeep==4.2.1  # For SOAP APIs (TRM)

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
